chore: Forgejo action to deploy site on webserver
This workflow requires an action runner that is directly running
on the webserver (here requested with the deployment-specific
`site-deploy` "runs-on tag".

The site is checked out using the DataLad clone action. Limited
functionality (for now), but good enough. In particular limiting
the clone depth of the submodules would be beneficial here (congo is
heavy).

We use a hugo github action that requires no dedicated container to make
things fast. The entire procedure takes less then a minute from action
start to deploy.